bashrc :: make sure the default variables are there
when a config is provided who does not contain all the configuration
options, everything should work fine. previously this was not the case,
if a file was provided, there could be non existing variables.

now this is fixed with a true default which is uses whenever a variable
is not defined.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>